final fixed
Fixed all the rubric notes. Kept snapcontainers but added some more photos. I like that the snap containers force the reader to consume specific parts of the story like when it was in print.

Added more notes to explain thought process and mark certain tags.

Fixed the use of col in the grid system.

Did not increase number of p tags too much because it disrupted the structure but I added a padded section which fixed the overlap on all breakpoints.

Used media query to remove scroll snap on smaller screens. Can be disorienting on things like phones.